move tests for reading to the parser test
